stripe landing pages styled, upgrade button into seperate component
