+++ src/client/pages/NotAuthorized.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { Button } from "../components/ui/Button";
-import Navbar from "../components/Navbar";
 
 const NotAuthorized = () => {
   return (
